wip: add user/role based authorization

+++ server/models/catModel.js
@@ -47,13 +47,13 @@ const addCat = async (cat, res) => {
   }
 };
 
-const deleteCatById = async (catId, res) => {
+const deleteCatById = async (catId, owner, res) => {
   console.log("model", catId);
   try {
     // TODO: do the LEFT (or INNER) JOIN to get owner's name as ownername (from wop_user table).
     const [rows] = await promisePool.query(
-      "DELETE FROM wop_cat WHERE cat_id = ?",
-      [catId]
+      "DELETE FROM wop_cat WHERE cat_id = ? AND owner = ?",
+      [catId, owner]
     );
     return rows;
   } catch (e) {
